backend/agents/guide/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai